Refactoring workspace page
