+++ src/common/hooks/index.ts
@@ -1,8 +1,8 @@
 
+export * from './useAddBlock'
+export * from './useDeleteBlock'
 
-export * from './useDelete'
 export * from './useGetAccountName'
-export * from './useGetSandboxData'
-export * from './useGetWorkspaceData'
-export * from './useGetTagsAndEntitiesData'
-export * from './useDeleteBlocks'
+
+export * from './sandbox'
+export * from './workspaceID'
